Refactor stream distribution package to streamplacement for better clarity and organization.
